refactoring due to some advices in the last PR
